+++ cub3D/src/draw_2.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   draw_2.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: paulo <paulo@student.42.fr>                +#+  +:+       +#+        */
+/*   By: pveiga-c <pveiga-c@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/31 09:33:22 by pveiga-c          #+#    #+#             */
-/*   Updated: 2024/02/14 12:40:32 by paulo            ###   ########.fr       */
+/*   Updated: 2024/02/19 15:10:55 by pveiga-c         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,8 +38,8 @@ void	my_pixel_put(t_image_data *img, int x, int y, int color)
 
 void	draw_wall(t_cub3d *cub3d, int pixel_w, int pixel_h)
 {
-	cub3d->tex.tex_y = abs((int)(((float)pixel_h - (float)SCREEN_HEIGHT / 2.0
-					+ (float)cub3d->tex.wall_line_h / 2.0) * cub3d->tex.step));
+	cub3d->tex.tex_y = abs((int)(((double)pixel_h - (double)SCREEN_HEIGHT / 2.0
+					+ (double)cub3d->tex.wall_line_h / 2.0) * cub3d->tex.step));
 	if (cub3d->ray.wall_side == 0 && cub3d->ray.ray_dir.x > 0)
 		cub3d->tex.color = get_pixel_color(&cub3d->east_view, cub3d->tex.tex_x,
 				cub3d->tex.tex_y);
